DL/Avanzado/VGG/requirements.txt
matplotlib==3.3.4
numpy==1.18.5
Pillow==7.2.0
pytorch-lightning==1.2.4
opencv-contrib-python==4.1.2.30
torch
torchvision